Move ImageClassificationModel to own file
